backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.7.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.19
pymupdf>=1.24.5
pytesseract>=0.3.10
Pillow>=10.3.0
paddleocr>=3.0.0
pandas>=2.2.2
openpyxl>=3.1.5
reportlab>=4.2.2
